signed/unsigned warnings fixed
these were noticed by the winbuilder service builds

+++ src/rcpp_sfs.cpp
@@ -54,8 +54,8 @@ sfs__matrix_to_DataFrame(SEXP E, const double zero_eps) {
     Rcpp::NumericMatrix matrix(E);
     Rcpp::IntegerVector rowidx,colidx;
     Rcpp::NumericVector value;
-    for(size_t row=0; row<matrix.nrow(); row++) {
-        for(size_t col=0; col<matrix.ncol(); col++) {
+    for(size_t row=0; row<(unsigned)matrix.nrow(); row++) {
+        for(size_t col=0; col<(unsigned)matrix.ncol(); col++) {
             double v = matrix(row,col);
             if(row!= col && fabs(v)>zero_eps) {
                 rowidx.push_back(row+1);
@@ -83,7 +83,7 @@ sfs__dist_to_DataFrame(SEXP R_dist, const double zero_eps) {
     // without diagonal elements. Typically rather dense; access via
     // LT_POS(size,col_idx,row_idx) (1-based)
     size_t length = LENGTH(R_dist);
-    int n = 1 + (int)sqrt(2*length);
+    size_t n = 1 + (int)sqrt(2*length);
     if(length != n*(n-1)/2) {
         throw std::runtime_error("dist object has invalid length");
     }
